requirements.txt
pytesseract==0.3.10
Pillow==10.0.1
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
